Extend supporting environment
* Abstract results generation and caching
* Trigger generation event
* Display default class name
