+++ src/server/server.py
@@ -53,6 +53,7 @@ import json
 import Config
 import Options
 import Database
+import Results
 import Language
 from Language import __
 
@@ -74,9 +75,6 @@ APP = Flask(__name__, static_url_path='/static')
 HEARTBEAT_THREAD = None
 HEARTBEAT_DATA_RATE_FACTOR = 5
 
-PILOT_ID_NONE = 0  # indicator value for no pilot configured
-HEAT_ID_NONE = 0  # indicator value for practice heat
-CLASS_ID_NONE = 0  # indicator value for unclassified heat
 FREQUENCY_ID_NONE = 0  # indicator value for node disabled
 
 ERROR_REPORT_INTERVAL_SECS = 600  # delay between comm-error reports to log
@@ -176,7 +174,7 @@ class Slave:
         pilot_id = Database.HeatNode.query.filter_by( \
             heat_id=RACE.current_heat, node_index=node_index).one_or_none().pilot_id
 
-        if pilot_id != PILOT_ID_NONE:
+        if pilot_id != Database.PILOT_ID_NONE:
 
             split_ts = data['timestamp'] + (PROGRAM_START_MILLIS_OFFSET - 1000.0*RACE.start_time_monotonic)
 
@@ -675,7 +673,7 @@ def api_heat(heat_id):
 
     payload = {
         'setup': heat,
-        'leaderboard': calc_leaderboard(heat_id=heat_id)
+        'leaderboard': Results.calc_leaderboard(DB, heat_id=heat_id)
     }
 
     return json.dumps({"heat": payload}, cls=AlchemyEncoder), 201, {'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*'}
@@ -728,12 +726,12 @@ def api_profile(profile_id):
 @APP.route('/api/race/current')
 def api_race_current():
     global RACE
-    if RACE.cacheStatus == CacheStatus.VALID:
+    if RACE.cacheStatus == Results.CacheStatus.VALID:
         results = RACE.results
     else:
-        results = calc_leaderboard(current_race=True)
+        results = Results.calc_leaderboard(DB, current_race=True)
         RACE.results = results
-        RACE.cacheStatus = CacheStatus.VALID
+        RACE.cacheStatus = Results.CacheStatus.VALID
 
     payload = {
         "raw_laps": RACE.node_laps,
@@ -754,7 +752,7 @@ def api_race_all():
 
     payload = {
         "heats": heats,
-        "leaderboard": calc_leaderboard()
+        "leaderboard": Results.calc_leaderboard(DB)
     }
 
     return json.dumps({"races": payload}, cls=AlchemyEncoder), 201, {'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*'}
@@ -797,7 +795,7 @@ def api_race(heat_id, round_id):
         'start_time_formatted': race.start_time_formatted,
         'nodes': pilotraces,
         'sort': Options.get('pilotSort'),
-        'leaderboard': calc_leaderboard(heat_id=heat_id, round_id=round_id)
+        'leaderboard': Results.calc_leaderboard(DB, heat_id=heat_id, round_id=round_id)
     }
 
     return json.dumps({"race": payload}, cls=AlchemyEncoder), 201, {'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*'}
@@ -1165,13 +1163,13 @@ def on_set_scan(data):
 @SOCKET_IO.on('add_heat')
 def on_add_heat():
     '''Adds the next available heat number to the database.'''
-    new_heat = Database.Heat(class_id=CLASS_ID_NONE, cacheStatus=CacheStatus.INVALID)
+    new_heat = Database.Heat(class_id=Database.CLASS_ID_NONE, cacheStatus=Results.CacheStatus.INVALID)
     DB.session.add(new_heat)
     DB.session.flush()
     DB.session.refresh(new_heat)
 
     for node in range(RACE.num_nodes): # Add next heat with empty pilots
-        DB.session.add(Database.HeatNode(heat_id=new_heat.id, node_index=node, pilot_id=PILOT_ID_NONE))
+        DB.session.add(Database.HeatNode(heat_id=new_heat.id, node_index=node, pilot_id=Database.PILOT_ID_NONE))
 
     DB.session.commit()
 
@@ -1207,7 +1205,7 @@ def on_alter_heat(data):
         for heatNode in Database.HeatNode.query.filter_by(heat_id=heat_id):
             RACE.node_pilots[heatNode.node_index] = heatNode.pilot_id
 
-            if heatNode.pilot_id is not PILOT_ID_NONE:
+            if heatNode.pilot_id is not Database.PILOT_ID_NONE:
                 RACE.node_teams[heatNode.node_index] = Database.Pilot.query.get(heatNode.pilot_id).team
             else:
                 RACE.node_teams[heatNode.node_index] = None
@@ -1255,7 +1253,7 @@ def on_delete_heat(data):
 @SOCKET_IO.on('add_race_class')
 def on_add_race_class():
     '''Adds the next available pilot id number in the database.'''
-    new_race_class = Database.RaceClass(name='New class', format_id=0, cacheStatus=CacheStatus.INVALID)
+    new_race_class = Database.RaceClass(name='New class', format_id=0, cacheStatus=Results.CacheStatus.INVALID)
     DB.session.add(new_race_class)
     DB.session.flush()
     DB.session.refresh(new_race_class)
@@ -1312,7 +1310,7 @@ def on_delete_class(data):
         DB.session.delete(race_class)
         for heat in Database.Heat.query.all():
             if heat.class_id == race_class.id:
-                heat.class_id = CLASS_ID_NONE
+                heat.class_id = Database.CLASS_ID_NONE
 
         DB.session.commit()
 
@@ -1367,7 +1365,7 @@ def on_alter_pilot(data):
     logger.info('Altered pilot {0} to {1}'.format(pilot_id, data))
     emit_pilot_data(noself=True) # Settings page, new pilot settings
     if 'callsign' in data:
-        invalidate_all_caches() # wipe caches (all have stored pilot names)
+        Results.invalidate_all_caches(DB) # wipe caches (all have stored pilot names)
         emit_round_data_notify() # live update rounds page
         emit_heat_data() # Settings page, new pilot callsign in heats
     if 'phonetic' in data:
@@ -1387,7 +1385,7 @@ def on_delete_pilot(data):
         DB.session.delete(pilot)
         for heatNode in Database.HeatNode.query.all():
             if heatNode.pilot_id == pilot.id:
-                heatNode.pilot_id = PILOT_ID_NONE
+                heatNode.pilot_id = Database.PILOT_ID_NONE
         DB.session.commit()
 
         logger.info('Pilot {0} deleted'.format(pilot.id))
@@ -1849,7 +1847,7 @@ def on_stage_race():
     heatNodes = Database.HeatNode.query.filter_by(heat_id=RACE.current_heat).all()
     for heatNode in heatNodes:
         if heatNode.node_index < RACE.num_nodes:
-            if heatNode.pilot_id != PILOT_ID_NONE:
+            if heatNode.pilot_id != Database.PILOT_ID_NONE:
                 valid_pilots = True
                 break
 
@@ -1865,7 +1863,7 @@ def on_stage_race():
         Events.trigger(Evt.RACE_STAGE)
         clear_laps() # Clear laps before race start
         init_node_cross_fields()  # set 'cur_pilot_id' and 'cross' fields on nodes
-        RACE.last_race_cacheStatus = CacheStatus.INVALID # invalidate last race results cache
+        RACE.last_race_cacheStatus = Results.CacheStatus.INVALID # invalidate last race results cache
         RACE.timer_running = False # indicate race timer not running
         RACE.race_status = RaceStatus.STAGING
         INTERFACE.set_race_status(RaceStatus.STAGING)
@@ -1921,7 +1919,7 @@ def findBestValues(node, node_index):
     current_class = heat.class_id
 
     # test for disabled node
-    if pilot is PILOT_ID_NONE or node.frequency is FREQUENCY_ID_NONE:
+    if pilot is Database.PILOT_ID_NONE or node.frequency is FREQUENCY_ID_NONE:
         logger.debug('Node {0} calibration: skipping disabled node'.format(node.index+1))
         return {
             'enter_at_level': node.enter_at_level,
@@ -1981,7 +1979,7 @@ def race_start_thread(start_token):
 
     # clear any lingering crossings at staging (if node rssi < enterAt)
     for node in INTERFACE.nodes:
-        if node.crossing_flag and node.frequency > 0 and node.current_pilot_id != PILOT_ID_NONE and \
+        if node.crossing_flag and node.frequency > 0 and node.current_pilot_id != Database.PILOT_ID_NONE and \
                     node.current_rssi < node.enter_at_level:
             logger.info("Forcing end crossing for node {0} at staging (rssi={1}, enterAt={2}, exitAt={3})".\
                        format(node.index+1, node.current_rssi, node.enter_at_level, node.exit_at_level))
@@ -2011,7 +2009,7 @@ def race_start_thread(start_token):
             node.history_times = []
             node.under_min_lap_count = 0
             # clear any lingering crossing (if rssi>enterAt then first crossing starts now)
-            if node.crossing_flag and node.frequency > 0 and node.current_pilot_id != PILOT_ID_NONE:
+            if node.crossing_flag and node.frequency > 0 and node.current_pilot_id != Database.PILOT_ID_NONE:
                 logger.info("Forcing end crossing for node {0} at start (rssi={1}, enterAt={2}, exitAt={3})".\
                            format(node.index+1, node.current_rssi, node.enter_at_level, node.exit_at_level))
                 INTERFACE.force_end_crossing(node.index)
@@ -2082,7 +2080,7 @@ def on_save_laps():
         format_id=Options.get('currentFormat'), \
         start_time = RACE.start_time_monotonic, \
         start_time_formatted = RACE.start_time.strftime("%Y-%m-%d %H:%M:%S"), \
-        cacheStatus=CacheStatus.INVALID
+        cacheStatus=Results.CacheStatus.INVALID
     )
     DB.session.add(new_race)
     DB.session.flush()
@@ -2093,7 +2091,7 @@ def on_save_laps():
             pilot_id = Database.HeatNode.query.filter_by( \
                 heat_id=RACE.current_heat, node_index=node_index).one().pilot_id
 
-            if pilot_id != PILOT_ID_NONE:
+            if pilot_id != Database.PILOT_ID_NONE:
                 new_pilotrace = Database.SavedPilotRace( \
                     race_id=new_race.id, \
                     node_index=node_index, \
@@ -2129,7 +2127,7 @@ def on_save_laps():
         'heat_id': RACE.current_heat,
         'round_id': new_race.round_id,
     }
-    gevent.spawn(build_race_results_caches, params)
+    gevent.spawn(Results.build_race_results_caches, DB, params)
 
     Events.trigger(Evt.LAPS_SAVE, {
         'race_id': new_race.id,
@@ -2189,7 +2187,7 @@ def on_resave_laps(data):
         'heat_id': heat_id,
         'round_id': round_id,
     }
-    gevent.spawn(build_race_results_caches, params)
+    gevent.spawn(Results.build_race_results_caches, DB, params)
 
     Events.trigger(Evt.LAPS_RESAVE, {
         'race_id': race_id,
@@ -2228,8 +2226,8 @@ def on_discard_laps(**kwargs):
 def clear_laps():
     '''Clear the current laps table.'''
     global RACE
-    RACE.last_race_results = calc_leaderboard(current_race=True)
-    RACE.last_race_cacheStatus = CacheStatus.VALID
+    RACE.last_race_results = Results.calc_leaderboard(DB, current_race=True)
+    RACE.last_race_cacheStatus = Results.CacheStatus.VALID
     RACE.laps_winner_name = None  # clear winner in first-to-X-laps race
     db_reset_current_laps() # Clear out the current laps table
     DB.session.query(Database.LapSplit).delete()
@@ -2242,7 +2240,7 @@ def init_node_cross_fields():
         heat_id=RACE.current_heat).all()
 
     for node in INTERFACE.nodes:
-        node.current_pilot_id = PILOT_ID_NONE
+        node.current_pilot_id = Database.PILOT_ID_NONE
         if node.frequency and node.frequency > 0:
             for heatnode in heatnodes:
                 if heatnode.node_index == node.index:
@@ -2281,7 +2279,7 @@ def on_set_current_heat(data):
     for heatNode in Database.HeatNode.query.filter_by(heat_id=new_heat_id):
         RACE.node_pilots[heatNode.node_index] = heatNode.pilot_id
 
-        if heatNode.pilot_id is not PILOT_ID_NONE:
+        if heatNode.pilot_id is not Database.PILOT_ID_NONE:
             RACE.node_teams[heatNode.node_index] = Database.Pilot.query.get(heatNode.pilot_id).team
         else:
             RACE.node_teams[heatNode.node_index] = None
@@ -2307,7 +2305,7 @@ def on_generate_heats(data):
     gevent.spawn(generate_heats, data)
 
 def generate_heats(data):
-    RESULTS_TIMEOUT = 5 # maximum time to wait for results to generate
+    RESULTS_TIMEOUT = 30 # maximum time to wait for results to generate
 
     '''Generate heats from qualifying class'''
     input_class = int(data['input_class'])
@@ -2316,7 +2314,7 @@ def generate_heats(data):
     pilots_per_heat = int(data['pilots_per_heat'])
     win_condition = data['win_condition']
 
-    if input_class == CLASS_ID_NONE:
+    if input_class == Database.CLASS_ID_NONE:
         results = {
             'by_race_time': []
         }
@@ -2334,7 +2332,7 @@ def generate_heats(data):
             results['by_race_time'].append(entry)
 
         win_condition = WinCondition.NONE
-        cacheStatus = CacheStatus.VALID
+        cacheStatus = Results.CacheStatus.VALID
     else:
         race_class = Database.RaceClass.query.get(input_class)
         race_format = Database.RaceFormat.query.get(race_class.format_id)
@@ -2344,18 +2342,23 @@ def generate_heats(data):
             cacheStatus = race_class.cacheStatus
         else:
             logger.error('Unable to fetch format from race class {0}'.format(input_class))
+            return False
 
-    if cacheStatus != CacheStatus.VALID:
-        pass
-        # build results ***
+    if cacheStatus == Results.CacheStatus.INVALID:
+        # build new results if needed
+        logger.info("No class cache available for {0}; regenerating".format(input_class))
+        race_class.cacheStatus = monotonic()
+        race_class.results = Results.calc_leaderboard(DB, class_id=race_class.id)
+        race_class.cacheStatus = Results.CacheStatus.VALID
+        DB.session.commit()
 
     time_now = monotonic()
     timeout = time_now + RESULTS_TIMEOUT
-    while cacheStatus != CacheStatus.VALID and time_now < timeout:
+    while cacheStatus != Results.CacheStatus.VALID and time_now < timeout:
         gevent.sleep()
         time_now = monotonic()
 
-    if cacheStatus == CacheStatus.VALID:
+    if cacheStatus == Results.CacheStatus.VALID:
         if win_condition == WinCondition.NONE:
             leaderboard = random.shuffle(results['by_race_time'])
         if win_condition == WinCondition.FASTEST_3_CONSECUTIVE:
@@ -2412,7 +2415,7 @@ def generate_heats(data):
         letters = __('ABCDEFGHIJKLMNOPQRSTUVWXYZ')
         for idx, heat in enumerate(reversed(generated_heats), start=1):
             ladder = letters[len(generated_heats) - idx]
-            new_heat = Database.Heat(class_id=output_class, cacheStatus=CacheStatus.INVALID, note=ladder + ' ' + suffix)
+            new_heat = Database.Heat(class_id=output_class, cacheStatus=Results.CacheStatus.INVALID, note=ladder + ' ' + suffix)
             DB.session.add(new_heat)
             DB.session.flush()
             DB.session.refresh(new_heat)
@@ -2421,15 +2424,19 @@ def generate_heats(data):
                 if node in heat:
                     DB.session.add(Database.HeatNode(heat_id=new_heat.id, node_index=node, pilot_id=heat[node]))
                 else:
-                    DB.session.add(Database.HeatNode(heat_id=new_heat.id, node_index=node, pilot_id=PILOT_ID_NONE))
+                    DB.session.add(Database.HeatNode(heat_id=new_heat.id, node_index=node, pilot_id=Database.PILOT_ID_NONE))
 
             DB.session.commit()
 
         logger.info("Generated {0} heats from class {1}".format(len(generated_heats), input_class))
+        SOCKET_IO.emit('heat_generate_done')
+
+        Events.trigger(Evt.HEAT_GENERATE)
 
         emit_heat_data()
     else:
         logger.warning("Unable to generate heats from class {0}: can't get valid results".format(input_class))
+        SOCKET_IO.emit('heat_generate_done')
 
 @SOCKET_IO.on('delete_lap')
 def on_delete_lap(data):
@@ -2768,7 +2775,7 @@ def emit_race_format(**params):
 def emit_current_laps(**params):
     '''Emits current laps.'''
     global RACE
-    if 'use_cache' in params and RACE.last_race_cacheStatus == CacheStatus.VALID:
+    if 'use_cache' in params and RACE.last_race_cacheStatus == Results.CacheStatus.VALID:
         emit_payload = RACE.last_race_laps
     else:
         current_laps = []
@@ -2883,7 +2890,7 @@ def emit_race_list(**params):
 
     '''
     heats_by_class = {}
-    heats_by_class[CLASS_ID_NONE] = [heat.id for heat in Database.Heat.query.filter_by(class_id=CLASS_ID_NONE).all()]
+    heats_by_class[Database.CLASS_ID_NONE] = [heat.id for heat in Database.Heat.query.filter_by(class_id=Database.CLASS_ID_NONE).all()]
     for race_class in Database.RaceClass.query.all():
         heats_by_class[race_class.id] = [heat.id for heat in Database.Heat.query.filter_by(class_id=race_class.id).all()]
 
@@ -2968,16 +2975,16 @@ def emit_round_data_thread(params, sid):
                             'node_index': pilotrace.node_index,
                             'laps': laps
                         })
-                    if round.cacheStatus == CacheStatus.INVALID:
-                        results = calc_leaderboard(heat_id=heat.heat_id, round_id=round.round_id)
+                    if round.cacheStatus == Results.CacheStatus.INVALID:
+                        results = Results.calc_leaderboard(DB, heat_id=heat.heat_id, round_id=round.round_id)
                         round.results = results
-                        round.cacheStatus = CacheStatus.VALID
+                        round.cacheStatus = Results.CacheStatus.VALID
                         DB.session.commit()
                     else:
                         checkStatus = True
                         while checkStatus:
                             gevent.idle()
-                            if round.cacheStatus == CacheStatus.VALID:
+                            if round.cacheStatus == Results.CacheStatus.VALID:
                                 results = round.results
                                 break
                             elif isinstance(round.cacheStatus, int) and round.cacheStatus < monotonic() + CACHE_TIMEOUT:
@@ -2990,16 +2997,16 @@ def emit_round_data_thread(params, sid):
                         'leaderboard': results
                     })
 
-                if heatdata.cacheStatus == CacheStatus.INVALID:
-                    results = calc_leaderboard(heat_id=heat.heat_id)
+                if heatdata.cacheStatus == Results.CacheStatus.INVALID:
+                    results = Results.calc_leaderboard(DB, heat_id=heat.heat_id)
                     heatdata.results = results
-                    heatdata.cacheStatus = CacheStatus.VALID
+                    heatdata.cacheStatus = Results.CacheStatus.VALID
                     DB.session.commit()
                 else:
                     checkStatus = True
                     while checkStatus:
                         gevent.idle()
-                        if heatdata.cacheStatus == CacheStatus.VALID:
+                        if heatdata.cacheStatus == Results.CacheStatus.VALID:
                             results = heatdata.results
                             break
                         elif isinstance(heatdata.cacheStatus, int) and heatdata.cacheStatus < monotonic() + CACHE_TIMEOUT:
@@ -3014,7 +3021,7 @@ def emit_round_data_thread(params, sid):
 
             gevent.sleep()
             heats_by_class = {}
-            heats_by_class[CLASS_ID_NONE] = [heat.id for heat in Database.Heat.query.filter_by(class_id=CLASS_ID_NONE).all()]
+            heats_by_class[Database.CLASS_ID_NONE] = [heat.id for heat in Database.Heat.query.filter_by(class_id=Database.CLASS_ID_NONE).all()]
             for race_class in Database.RaceClass.query.all():
                 heats_by_class[race_class.id] = [heat.id for heat in Database.Heat.query.filter_by(class_id=race_class.id).all()]
 
@@ -3022,16 +3029,16 @@ def emit_round_data_thread(params, sid):
             current_classes = {}
             for race_class in Database.RaceClass.query.all():
 
-                if race_class.cacheStatus == CacheStatus.INVALID:
-                    results = calc_leaderboard(class_id=race_class.id)
+                if race_class.cacheStatus == Results.CacheStatus.INVALID:
+                    results = Results.calc_leaderboard(DB, class_id=race_class.id)
                     race_class.results = results
-                    race_class.cacheStatus = CacheStatus.VALID
+                    race_class.cacheStatus = Results.CacheStatus.VALID
                     DB.session.commit()
                 else:
                     checkStatus = True
                     while checkStatus:
                         gevent.idle()
-                        if race_class.cacheStatus == CacheStatus.VALID:
+                        if race_class.cacheStatus == Results.CacheStatus.VALID:
                             results = race_class.results
                             break
                         elif isinstance(race_class.cacheStatus, int) and race_class.cacheStatus < monotonic() + CACHE_TIMEOUT:
@@ -3046,17 +3053,17 @@ def emit_round_data_thread(params, sid):
 
             gevent.sleep()
 
-            if Options.get("eventResults_cacheStatus") == CacheStatus.INVALID:
-                results = calc_leaderboard()
+            if Options.get("eventResults_cacheStatus") == Results.CacheStatus.INVALID:
+                results = Results.calc_leaderboard(DB)
                 Options.set("eventResults", json.dumps(results))
-                Options.set("eventResults_cacheStatus", CacheStatus.VALID)
+                Options.set("eventResults_cacheStatus", Results.CacheStatus.VALID)
                 DB.session.commit()
             else:
                 checkStatus = True
                 while checkStatus:
                     gevent.idle()
                     status = Options.get("eventResults_cacheStatus")
-                    if status == CacheStatus.VALID:
+                    if status == Results.CacheStatus.VALID:
                         results = json.loads(Options.get("eventResults"))
                         break
                     elif isinstance(status, int) and status < monotonic() + CACHE_TIMEOUT:
@@ -3080,503 +3087,22 @@ def emit_round_data_thread(params, sid):
         else:
             SOCKET_IO.emit('round_data', emit_payload, namespace='/')
 
-'''
-Results generation and caching
-'''
-
-class CacheStatus:
-    INVALID = 'invalid'
-    VALID = 'valid'
-
-@SOCKET_IO.on('wipe_cache')
-def invalidate_all_caches():
-    ''' Check all caches and invalidate any paused builds '''
-    for race in Database.SavedRaceMeta.query.all():
-        race.cacheStatus = CacheStatus.INVALID
-
-    for heat in Database.Heat.query.all():
-        heat.cacheStatus = CacheStatus.INVALID
-
-    for race_class in Database.RaceClass.query.all():
-        race_class.cacheStatus = CacheStatus.INVALID
-
-    DB.session.commit()
-
-    Options.set("eventResults_cacheStatus", CacheStatus.INVALID)
-
-    global FULL_RESULTS_CACHE_VALID
-    FULL_RESULTS_CACHE_VALID = False
-
-    Events.trigger(Evt.CACHE_CLEAR)
-
-    logger.info('All Result caches invalidated')
-
-def normalize_cache_status():
-    ''' Check all caches and invalidate any paused builds '''
-    for race in Database.SavedRaceMeta.query.all():
-        if race.cacheStatus != CacheStatus.VALID:
-            race.cacheStatus = CacheStatus.INVALID
-
-    for heat in Database.Heat.query.all():
-        if heat.cacheStatus != CacheStatus.VALID:
-            heat.cacheStatus = CacheStatus.INVALID
-
-    for race_class in Database.RaceClass.query.all():
-        if race_class.cacheStatus != CacheStatus.VALID:
-            race_class.cacheStatus = CacheStatus.INVALID
-
-    if Options.get("eventResults_cacheStatus") != CacheStatus.VALID:
-        Options.set("eventResults_cacheStatus", CacheStatus.INVALID)
-
-    DB.session.commit()
-
-    global FULL_RESULTS_CACHE_VALID
-    FULL_RESULTS_CACHE_VALID = False
-
-    logger.info('All Result caches normalized')
-
-def build_race_results_caches(params):
-    global FULL_RESULTS_CACHE
-    FULL_RESULTS_CACHE = False
-    token = monotonic()
-
-    race = Database.SavedRaceMeta.query.get(params['race_id'])
-    heat = Database.Heat.query.get(params['heat_id'])
-    if heat.class_id != CLASS_ID_NONE:
-        race_class = Database.RaceClass.query.get(heat.class_id)
-
-    race.cacheStatus = token
-    heat.cacheStatus = token
-    if heat.class_id != CLASS_ID_NONE:
-        race_class.cacheStatus = token
-    Options.set("eventResults_cacheStatus", token)
-    DB.session.commit()
-
-    # rebuild race result
-    gevent.sleep()
-    if race.cacheStatus == token:
-        race.results = calc_leaderboard(heat_id=params['heat_id'], round_id=params['round_id'])
-        race.cacheStatus = CacheStatus.VALID
-        DB.session.commit()
-
-    # rebuild heat summary
-    gevent.sleep()
-    if heat.cacheStatus == token:
-        heat.results = calc_leaderboard(heat_id=params['heat_id'])
-        heat.cacheStatus = CacheStatus.VALID
-        DB.session.commit()
-
-    # rebuild class summary
-    if heat.class_id != CLASS_ID_NONE:
-        if race_class.cacheStatus == token:
-            gevent.sleep()
-            race_class.results = calc_leaderboard(class_id=heat.class_id)
-            race_class.cacheStatus = CacheStatus.VALID
-            DB.session.commit()
-
-    # rebuild event summary
-    gevent.sleep()
-    Options.set("eventResults", json.dumps(calc_leaderboard()))
-    Options.set("eventResults_cacheStatus", CacheStatus.VALID)
-
-    logger.info('Built result caches: Race {0}, Heat {1}, Class {2}, Event'.format(params['race_id'], params['heat_id'], heat.class_id))
-
-def calc_leaderboard(**params):
-    ''' Generates leaderboards '''
-    USE_CURRENT = False
-    USE_ROUND = None
-    USE_HEAT = None
-    USE_CLASS = None
-
-    if ('current_race' in params):
-        USE_CURRENT = True
-
-    if ('class_id' in params):
-        USE_CLASS = params['class_id']
-    elif ('round_id' in params and 'heat_id' in params):
-        USE_ROUND = params['round_id']
-        USE_HEAT = params['heat_id']
-    elif ('heat_id' in params):
-        USE_ROUND = None
-        USE_HEAT = params['heat_id']
-
-    # Get profile (current), frequencies (current), race query (saved), and race format (all)
-    if USE_CURRENT:
-        profile = getCurrentProfile()
-        profile_freqs = json.loads(profile.frequencies)
-        race_format = getCurrentRaceFormat()
-    else:
-        if USE_CLASS:
-            race_query = Database.SavedRaceMeta.query.filter_by(class_id=USE_CLASS)
-            if race_query.count() >= 1:
-                current_format = Database.RaceClass.query.get(USE_CLASS).format_id
-            else:
-                current_format = None
-        elif USE_HEAT:
-            if USE_ROUND:
-                race_query = Database.SavedRaceMeta.query.filter_by(heat_id=USE_HEAT, round_id=USE_ROUND)
-                current_format = race_query.first().format_id
-            else:
-                race_query = Database.SavedRaceMeta.query.filter_by(heat_id=USE_HEAT)
-                if race_query.count() >= 1:
-                    heat_class = race_query.first().class_id
-                    if heat_class:
-                        current_format = Database.RaceClass.query.get(heat_class).format_id
-                    else:
-                        current_format = None
-                else:
-                    current_format = None
-        else:
-            race_query = Database.SavedRaceMeta.query
-            current_format = None
-
-        selected_races = race_query.all()
-        racelist = [r.id for r in selected_races]
-
-        if current_format:
-            race_format = Database.RaceFormat.query.get(current_format)
-        else:
-            race_format = None
-
-    gevent.sleep()
-    # Get the pilot ids for all relevant races
-    # Add pilot callsigns
-    # Add pilot team names
-    # Get total laps for each pilot
-    # Get hole shot laps
-    pilot_ids = []
-    callsigns = []
-    nodes = []
-    team_names = []
-    max_laps = []
-    current_laps = []
-    holeshots = []
-
-    for pilot in Database.Pilot.query.filter(Database.Pilot.id != PILOT_ID_NONE):
-        gevent.sleep()
-        if USE_CURRENT:
-            laps = []
-            for node_index in RACE.node_pilots:
-                if RACE.node_pilots[node_index] == pilot.id and node_index < RACE.num_nodes:
-                    laps = RACE.get_active_laps()[node_index]
-                    break
-
-            if laps:
-                max_lap = len(laps) - 1
-            else:
-                max_lap = 0
-
-            current_heat = Database.HeatNode.query.filter_by(heat_id=RACE.current_heat, pilot_id=pilot.id).first()
-            if current_heat and profile_freqs["f"][current_heat.node_index] != FREQUENCY_ID_NONE:
-                pilot_ids.append(pilot.id)
-                callsigns.append(pilot.callsign)
-                nodes.append(current_heat.node_index)
-                team_names.append(pilot.team)
-                max_laps.append(max_lap)
-                current_laps.append(laps)
-        else:
-            # find hole shots
-            holeshot_laps = []
-            for race in racelist:
-                pilotraces = Database.SavedPilotRace.query \
-                    .filter(Database.SavedPilotRace.pilot_id == pilot.id, \
-                    Database.SavedPilotRace.race_id == race \
-                    ).all()
-
-                for pilotrace in pilotraces:
-                    gevent.sleep()
-                    holeshot_lap = Database.SavedRaceLap.query \
-                        .filter(Database.SavedRaceLap.pilotrace_id == pilotrace.id, \
-                            Database.SavedRaceLap.deleted != 1, \
-                            ).order_by(Database.SavedRaceLap.lap_time_stamp).first()
-
-                    if holeshot_lap:
-                        holeshot_laps.append(holeshot_lap.id)
-
-            # get total laps
-            stat_query = DB.session.query(DB.func.count(Database.SavedRaceLap.id)) \
-                .filter(Database.SavedRaceLap.pilot_id == pilot.id, \
-                    Database.SavedRaceLap.deleted != 1, \
-                    Database.SavedRaceLap.race_id.in_(racelist), \
-                    ~Database.SavedRaceLap.id.in_(holeshot_laps))
-
-            max_lap = stat_query.scalar()
-            if max_lap > 0:
-                pilot_ids.append(pilot.id)
-                callsigns.append(pilot.callsign)
-                nodes.append(holeshot_lap.node_index)
-                team_names.append(pilot.team)
-                max_laps.append(max_lap)
-                holeshots.append(holeshot_laps)
-
-    total_time = []
-    last_lap = []
-    average_lap = []
-    fastest_lap = []
-    consecutives = []
-    fastest_lap_source = []
-    consecutives_source = []
-
-    for i, pilot in enumerate(pilot_ids):
-        gevent.sleep()
-        # Get the total race time for each pilot
-        if max_laps[i] is 0:
-            total_time.append(0) # Add zero if no laps completed
-        else:
-            if USE_CURRENT:
-                race_total = 0
-                for lap in current_laps[i]:
-                    race_total += lap['lap_time']
-
-                total_time.append(race_total)
-
-            else:
-                stat_query = DB.session.query(DB.func.sum(Database.SavedRaceLap.lap_time)) \
-                    .filter(Database.SavedRaceLap.pilot_id == pilot, \
-                        Database.SavedRaceLap.deleted != 1, \
-                        Database.SavedRaceLap.race_id.in_(racelist))
-
-                total_time.append(stat_query.scalar())
-
-        gevent.sleep()
-        # Get the last lap for each pilot (current race only)
-        if max_laps[i] is 0:
-            last_lap.append(None) # Add zero if no laps completed
-        else:
-            if USE_CURRENT:
-                last_lap.append(current_laps[i][-1]['lap_time'])
-            else:
-                last_lap.append(None)
-
-        gevent.sleep()
-        # Get the average lap time for each pilot
-        if max_laps[i] is 0:
-            average_lap.append(0) # Add zero if no laps completed
-        else:
-            if USE_CURRENT:
-                avg_lap = (current_laps[i][-1]['lap_time_stamp'] - current_laps[i][0]['lap_time_stamp']) / (len(current_laps[i]) - 1)
-
-                '''
-                timed_laps = filter(lambda x : x['lap_number'] > 0, current_laps[i])
-
-                lap_total = 0
-                for lap in timed_laps:
-                    lap_total += lap['lap_time']
-
-                avg_lap = lap_total / len(timed_laps)
-                '''
-
-            else:
-                stat_query = DB.session.query(DB.func.avg(Database.SavedRaceLap.lap_time)) \
-                    .filter(Database.SavedRaceLap.pilot_id == pilot, \
-                        Database.SavedRaceLap.deleted != 1, \
-                        Database.SavedRaceLap.race_id.in_(racelist), \
-                        ~Database.SavedRaceLap.id.in_(holeshots[i]))
-
-                avg_lap = stat_query.scalar()
-
-            average_lap.append(avg_lap)
-
-        gevent.sleep()
-        # Get the fastest lap time for each pilot
-        if max_laps[i] is 0:
-            fastest_lap.append(0) # Add zero if no laps completed
-            fastest_lap_source.append(None)
-        else:
-            if USE_CURRENT:
-                timed_laps = filter(lambda x : x['lap_number'] > 0, current_laps[i])
-
-                fast_lap = sorted(timed_laps, key=lambda val : val['lap_time'])[0]['lap_time']
-                fastest_lap_source.append(None)
-            else:
-                stat_query = DB.session.query(DB.func.min(Database.SavedRaceLap.lap_time).label('time'), Database.SavedRaceLap.race_id) \
-                    .filter(Database.SavedRaceLap.pilot_id == pilot, \
-                        Database.SavedRaceLap.deleted != 1, \
-                        Database.SavedRaceLap.race_id.in_(racelist), \
-                        ~Database.SavedRaceLap.id.in_(holeshots[i])).one()
-
-                fast_lap = stat_query.time
-
-                if USE_HEAT:
-                    fastest_lap_source.append(None)
-                else:
-                    source_query = Database.SavedRaceMeta.query.get(stat_query.race_id)
-                    fast_lap_round = source_query.round_id
-                    fast_lap_heat = source_query.heat_id
-                    fast_lap_heatnote = Database.Heat.query.get(fast_lap_heat).note
-
-                    if fast_lap_heatnote:
-                        source_text = fast_lap_heatnote + ' / ' + __('Round') + ' ' + str(fast_lap_round)
-                    else:
-                        source_text = __('Heat') + ' ' + str(fast_lap_heat) + ' / ' + __('Round') + ' ' + str(fast_lap_round)
-
-                    fastest_lap_source.append(source_text)
-
-            fastest_lap.append(fast_lap)
-
-        gevent.sleep()
-        # find best consecutive 3 laps
-        if max_laps[i] < 3:
-            consecutives.append(None)
-            consecutives_source.append(None)
-        else:
-            all_consecutives = []
-
-            if USE_CURRENT:
-                thisrace = current_laps[i][1:]
-
-                for j in range(len(thisrace) - 2):
-                    gevent.sleep()
-                    all_consecutives.append({
-                        'time': thisrace[j]['lap_time'] + thisrace[j+1]['lap_time'] + thisrace[j+2]['lap_time'],
-                        'race_id': None,
-                    })
-
-            else:
-                for race_id in racelist:
-                    gevent.sleep()
-                    thisrace = DB.session.query(Database.SavedRaceLap.lap_time) \
-                        .filter(Database.SavedRaceLap.pilot_id == pilot, \
-                            Database.SavedRaceLap.race_id == race_id, \
-                            Database.SavedRaceLap.deleted != 1, \
-                            ~Database.SavedRaceLap.id.in_(holeshots[i]) \
-                            ).all()
-
-                    if len(thisrace) >= 3:
-                        for j in range(len(thisrace) - 2):
-                            gevent.sleep()
-                            all_consecutives.append({
-                                'time': thisrace[j].lap_time + thisrace[j+1].lap_time + thisrace[j+2].lap_time,
-                                'race_id': race_id
-                            })
-
-            # Sort consecutives
-            all_consecutives.sort(key = lambda x: (x['time'] is None, x['time']))
-            # Get lowest not-none value (if any)
-
-            if all_consecutives:
-                consecutives.append(all_consecutives[0]['time'])
-
-                if USE_CURRENT:
-                    consecutives_source.append(None)
-                else:
-                    source_query = Database.SavedRaceMeta.query.get(all_consecutives[0]['race_id'])
-                    if source_query:
-                        fast_lap_round = source_query.round_id
-                        fast_lap_heat = source_query.heat_id
-                        fast_lap_heatnote = Database.Heat.query.get(fast_lap_heat).note
-
-                        if fast_lap_heatnote:
-                            source_text = fast_lap_heatnote + ' / ' + __('Round') + ' ' + str(fast_lap_round)
-                        else:
-                            source_text = __('Heat') + ' ' + str(fast_lap_heat) + ' / ' + __('Round') + ' ' + str(fast_lap_round)
-
-                        consecutives_source.append(source_text)
-                    else:
-                        consecutives_source.append(None)
-
-            else:
-                consecutives.append(None)
-                consecutives_source.append(None)
-
-    gevent.sleep()
-    # Combine for sorting
-    leaderboard = zip(callsigns, max_laps, total_time, average_lap, fastest_lap, team_names, consecutives, fastest_lap_source, consecutives_source, last_lap, pilot_ids, nodes)
-
-    # Reverse sort max_laps x[1], then sort on total time x[2]
-    leaderboard_by_race_time = sorted(leaderboard, key = lambda x: (-x[1], x[2]))
-
-    leaderboard_total_data = []
-    for i, row in enumerate(leaderboard_by_race_time, start=1):
-        leaderboard_total_data.append({
-            'position': i,
-            'callsign': row[0],
-            'laps': row[1],
-            'behind': (leaderboard_by_race_time[0][1] - row[1]),
-            'total_time': time_format(row[2]),
-            'average_lap': time_format(row[3]),
-            'fastest_lap': time_format(row[4]),
-            'team_name': row[5],
-            'consecutives': time_format(row[6]),
-            'fastest_lap_source': row[7],
-            'consecutives_source': row[8],
-            'last_lap': row[9],
-            'pilot_id': row[10],
-            'node': row[11],
-        })
-
-    gevent.sleep()
-    # Sort fastest_laps x[4]
-    leaderboard_by_fastest_lap = sorted(leaderboard, key = lambda x: (x[4] if x[4] > 0 else float('inf')))
-
-    leaderboard_fast_lap_data = []
-    for i, row in enumerate(leaderboard_by_fastest_lap, start=1):
-        leaderboard_fast_lap_data.append({
-            'position': i,
-            'callsign': row[0],
-            'total_time': time_format(row[2]),
-            'average_lap': time_format(row[3]),
-            'fastest_lap': time_format(row[4]),
-            'team_name': row[5],
-            'consecutives': time_format(row[6]),
-            'fastest_lap_source': row[7],
-            'consecutives_source': row[8],
-            'last_lap': row[9],
-            'pilot_id': row[10],
-            'node': row[11],
-        })
-
-    gevent.sleep()
-    # Sort consecutives x[6]
-    leaderboard_by_consecutives = sorted(leaderboard, key = lambda x: (x[6] if x[6] > 0 else float('inf')))
-
-    leaderboard_consecutives_data = []
-    for i, row in enumerate(leaderboard_by_consecutives, start=1):
-        leaderboard_consecutives_data.append({
-            'position': i,
-            'callsign': row[0],
-            'total_time': time_format(row[2]),
-            'average_lap': time_format(row[3]),
-            'fastest_lap': time_format(row[4]),
-            'team_name': row[5],
-            'consecutives': time_format(row[6]),
-            'fastest_lap_source': row[7],
-            'consecutives_source': row[8],
-            'last_lap': row[9],
-            'pilot_id': row[10],
-            'node': row[11],
-        })
-
-    leaderboard_output = {
-        'by_race_time': leaderboard_total_data,
-        'by_fastest_lap': leaderboard_fast_lap_data,
-        'by_consecutives': leaderboard_consecutives_data
-    }
-
-    if race_format:
-        leaderboard_output['meta'] = {
-            'win_condition': race_format.win_condition,
-            'team_racing_mode': race_format.team_racing_mode,
-        }
-    else:
-        leaderboard_output['meta'] = {
-            'win_condition': WinCondition.NONE,
-            'team_racing_mode': False
-        }
-
-    return leaderboard_output
+@SOCKET_IO.on('clean_cache')
+def clean_results_cache():
+    ''' expose for frontend debugging '''
+    Results.invalidate_all_caches(DB)
 
 def emit_current_leaderboard(**params):
     '''Emits leaderboard.'''
     global RACE
-    if 'use_cache' in params and RACE.last_race_cacheStatus == CacheStatus.VALID:
+    if 'use_cache' in params and RACE.last_race_cacheStatus == Results.CacheStatus.VALID:
         emit_payload = RACE.last_race_results
-    elif RACE.cacheStatus == CacheStatus.VALID:
+    elif RACE.cacheStatus == Results.CacheStatus.VALID:
         emit_payload = RACE.results
     else:
-        results = calc_leaderboard(current_race=True)
+        results = Results.calc_leaderboard(DB, current_race=True)
         RACE.results = results
-        RACE.cacheStatus = CacheStatus.VALID
+        RACE.cacheStatus = Results.CacheStatus.VALID
         emit_payload = results
 
     if ('nobroadcast' in params):
@@ -3730,7 +3256,7 @@ def emit_current_heat(**params):
     callsigns = []
                              # dict for current heat with key=node_index, value=pilot_id
     node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-        filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+        filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
     for node_index in range(RACE.num_nodes):
         pilot_id = node_pilot_dict.get(node_index)
@@ -3748,7 +3274,7 @@ def emit_current_heat(**params):
     heat_note = heat_data.note
 
     heat_format = None
-    if heat_data.class_id != CLASS_ID_NONE:
+    if heat_data.class_id != Database.CLASS_ID_NONE:
         heat_format = Database.RaceClass.query.get(heat_data.class_id).format_id
 
     emit_payload = {
@@ -3769,7 +3295,7 @@ def get_team_laps_info(cur_pilot_id=-1, num_laps_win=0):
     profile_freqs = json.loads(getCurrentProfile().frequencies)
                              # dict for current heat with key=node_index, value=pilot_id
     node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-                      filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+                      filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
     for node in INTERFACE.nodes:
         if profile_freqs["f"][node.index] != FREQUENCY_ID_NONE:
@@ -3848,7 +3374,7 @@ def check_pilot_laps_win(pass_node_index, num_laps_win):
     profile_freqs = json.loads(getCurrentProfile().frequencies)
                              # dict for current heat with key=node_index, value=pilot_id
     node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-                      filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+                      filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
     for node in INTERFACE.nodes:
         if profile_freqs["f"][node.index] != FREQUENCY_ID_NONE:
@@ -3879,7 +3405,7 @@ def check_team_laps_win(t_laps_dict, num_laps_win, pilot_team_dict, pass_node_in
         profile_freqs = None
                                   # dict for current heat with key=node_index, value=pilot_id
         node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-                          filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+                          filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
         for node in INTERFACE.nodes:  # check if (other) pilot node is crossing gate
             if node.crossing_flag and node.index != pass_node_index:
@@ -3967,7 +3493,7 @@ def check_most_laps_win(pass_node_index=-1, t_laps_dict=None, pilot_team_dict=No
                             if profile_freqs["f"][node.index] != FREQUENCY_ID_NONE:  # node is enabled
                                 if not node_pilot_dict:
                                     node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-                                              filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+                                              filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
                                 pilot_id = node_pilot_dict.get(node.index)
                                 if pilot_id:  # node has pilot assigned to it
@@ -4004,7 +3530,7 @@ def check_most_laps_win(pass_node_index=-1, t_laps_dict=None, pilot_team_dict=No
         profile_freqs = json.loads(getCurrentProfile().frequencies)
                                   # dict for current heat with key=node_index, value=pilot_id
         node_pilot_dict = dict(Database.HeatNode.query.with_entities(Database.HeatNode.node_index, Database.HeatNode.pilot_id). \
-                          filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=PILOT_ID_NONE).all())
+                          filter(Database.HeatNode.heat_id==RACE.current_heat, Database.HeatNode.pilot_id!=Database.PILOT_ID_NONE).all())
 
         for node in INTERFACE.nodes:  # load per-pilot data into 'pilots_list'
             if profile_freqs["f"][node.index] != FREQUENCY_ID_NONE:
@@ -4388,7 +3914,7 @@ def pass_record_callback(node, lap_timestamp_absolute, source):
                 heat_id=RACE.current_heat, node_index=node.index).one().pilot_id
 
             # reject passes before race start and with disabled (no-pilot) nodes
-            if pilot_id != PILOT_ID_NONE:
+            if pilot_id != Database.PILOT_ID_NONE:
                 if lap_timestamp_absolute >= RACE.start_time_monotonic:
 
                     lap_time_stamp = (lap_timestamp_absolute - RACE.start_time_monotonic)
@@ -4437,8 +3963,8 @@ def pass_record_callback(node, lap_timestamp_absolute, source):
                             'deleted': False
                         })
 
-                        RACE.results = calc_leaderboard(current_race=True)
-                        RACE.cacheStatus = CacheStatus.VALID
+                        RACE.results = Results.calc_leaderboard(DB, current_race=True)
+                        RACE.cacheStatus = Results.CacheStatus.VALID
 
                         Events.trigger(Evt.RACE_LAP_RECORDED, {
                             'node': node.index,
@@ -4555,7 +4081,7 @@ def node_crossing_callback(node):
 
     if led_manager.isEnabled() and RACE.race_status == RaceStatus.RACING:  # if race is in progress
         # if pilot assigned to node and first crossing is complete
-        if node.current_pilot_id != PILOT_ID_NONE and node.first_cross_flag:
+        if node.current_pilot_id != Database.PILOT_ID_NONE and node.first_cross_flag:
             # first crossing has happened; if 'enter' then show indicator,
             #  if first event is 'exit' then ignore (because will be end of first crossing)
             if node.crossing_flag:
@@ -4651,7 +4177,7 @@ def db_reset_current_laps():
     for idx in range(RACE.num_nodes):
         RACE.node_laps[idx] = []
 
-    RACE.cacheStatus = CacheStatus.INVALID
+    RACE.cacheStatus = Results.CacheStatus.INVALID
     logger.info('Database current laps reset')
 
 def db_reset_saved_races():
@@ -4792,7 +4318,7 @@ def db_reset_options_defaults():
     Options.set("colorNode_6", "#3fff3f")
     Options.set("colorNode_7", "#00bfff")
     # Event results cache
-    Options.set("eventResults_cacheStatus", CacheStatus.INVALID)
+    Options.set("eventResults_cacheStatus", Results.CacheStatus.INVALID)
 
     logger.info("Reset global settings")
 
@@ -4937,7 +4463,7 @@ def expand_heats():
         for node in range(RACE.num_nodes):
             heat_row = Database.HeatNode.query.filter_by(heat_id=heat_ids.id, node_index=node)
             if not heat_row.count():
-                DB.session.add(Database.HeatNode(heat_id=heat_ids.id, node_index=node, pilot_id=PILOT_ID_NONE))
+                DB.session.add(Database.HeatNode(heat_id=heat_ids.id, node_index=node, pilot_id=Database.PILOT_ID_NONE))
 
     DB.session.commit()
 
@@ -5114,13 +4640,13 @@ if Database.Heat.query.first():
     for heatNode in Database.HeatNode.query.filter_by(heat_id=RACE.current_heat):
         RACE.node_pilots[heatNode.node_index] = heatNode.pilot_id
 
-        if heatNode.pilot_id is not PILOT_ID_NONE:
+        if heatNode.pilot_id is not Database.PILOT_ID_NONE:
             RACE.node_teams[heatNode.node_index] = Database.Pilot.query.get(heatNode.pilot_id).team
         else:
             RACE.node_teams[heatNode.node_index] = None
 
 # Normalize results caches
-normalize_cache_status()
+Results.normalize_cache_status(DB)
 
 # Create LED object with appropriate configuration
 strip = None
