+++ src/server/templates/settings.html
@@ -588,9 +588,13 @@
 				pilots_per_heat: parseInt($('#generator_pilots_per_heat').val()),
 				win_condition: parseInt($('#generator_win_condition').val()),
 			};
-			// $('button#generate_heats').prop('disabled', true);
+			$('button#generate_heats').prop('disabled', true);
 			socket.emit('generate_heats', data);
 			return false;
+		});
+
+		socket.on('heat_generate_done', function (msg) {
+			$('button#generate_heats').prop('disabled', false);
 		});
 
 		/* Classes */
@@ -632,8 +636,14 @@
 
 					el.appendTo(classlist);
 
-					$('#generator_input_class').append($('<option value="' + msg.classes[i].id + '">' + msg.classes[i].name + '</option>'));
-					$('#generator_output_class').append($('<option value="' + msg.classes[i].id + '">' + msg.classes[i].name + '</option>'));
+					if (race_class.name) {
+						$('#generator_input_class').append($('<option value="' + race_class.id + '">' + __('Name') + ' (' + __('Class') + ' ' + race_class.id + '</option>'));
+						$('#generator_output_class').append($('<option value="' + race_class.id + '">' + __('Name') + ' (' + __('Class') + ' ' + race_class.id + '</option>'));
+					} else {
+						$('#generator_input_class').append($('<option value="' + race_class.id + '">' + __('Name') + ' (' + __('Class') + ' ' + race_class.id + '</option>'));
+						$('#generator_output_class').append($('<option value="' + race_class.id + '">' + __('Name') + ' (' + __('Class') + ' ' + race_class.id + '</option>'));
+					}
+
 				}
 				classlist.appendTo($('.race_classes'));
 			} else {
